requirements.txt
pynput>=1.7.7
Pillow>=11.0.0
